Fix deadlock in Rust Future::get
using future.get() would deadlock since the lock was still held by the
parked thread and when the set_ok call came to wake up that thread, it
would block on the lock waiting for the thread it was supposed to unpark
to unlock it.

This fixes the bug by properly unlocking the lock before parking.

Change-Id: I2ac7293f196a8964cabce6660a9b0a8ee0510a41