Don't close handles in the HandlePeer finalizer.
Finalizers aren't called in a predictable order. If a handle object is
finalized before a waiter that's waiting on that handle, cancelling the
waiter will (as of recent changes to the message loop) fail and abort
the process.

All handles are closed in the isolate cleanup callback (via the
DartState destructor) so we don't leak handles.

Change-Id: Ic997956acd9220ff349aa777efb9d8b8871f94be