Update Dart FIDL bindings.
Instead of representing handles to Dart as integers (tracked in the
handle table) use Handle (from dart:fidl.internal).

Handle waiters are now one-shot and owned by the Handle that they're
waiting for.

There are also clean-ups in the Dart code to make it pass analysis
cleanly. The analyzer was key in ensuring the refactor was valid so
that clean-up was important.

Change-Id: I43437b9ec7b215288c1b2f73732cc30b69c1cf66